requirements.txt
flask>=2.3.0
flask-httpauth>=4.8.0
opencv-python>=4.8.0
numpy>=1.24.0
RPi.GPIO>=0.7.1  # Tylko dla Raspberry Pi
